+++ src/sif/nocost.cc
@@ -213,7 +213,10 @@ public:
   }
 };
 
-void ParseNoCostOptions(const rapidjson::Document&, const std::string&, Costing* c) {
+void ParseNoCostOptions(const rapidjson::Document&,
+                        const std::string& /*costing_options_key*/,
+                        Costing* c,
+                        google::protobuf::RepeatedPtrField<CodedDescription>& /*warnings*/) {
   // this is probably not needed but its part of the contract for costing..
   c->set_type(Costing::none_);
   c->set_name(Costing_Enum_Name(c->type()));