+++ src/thor/matrix_action.cc
@@ -6,6 +6,7 @@
 #include "thor/timedistancematrix.h"
 #include "thor/worker.h"
 #include "tyr/serializers.h"
+#include <valhalla/worker.h>
 
 using namespace valhalla;
 using namespace valhalla::tyr;
@@ -126,6 +127,11 @@ std::string thor_worker_t::matrix(Api& request) {
     algo->SourceToTarget(request, *reader, mode_costing, mode,
                          max_matrix_distance.find(costing)->second);
     return tyr::serializeMatrix(request);
+  }
+
+  // no matrix_locations for CostMatrix
+  if (options.matrix_locations() != std::numeric_limits<uint32_t>::max()) {
+    add_warning(request, 211);
   }
 
   // for costmatrix try a second pass if the first didn't work out