+++ test/gurka/test_fixed_speed.cc
@@ -70,15 +70,20 @@ TEST_F(FixedSpeedTest, MotorscooterFixedSpeed) {
 }
 
 TEST_F(FixedSpeedTest, ClampMaxSpeed) {
-  Options options;
+  Api api;
+  auto& options = *api.mutable_options();
   rapidjson::Document dom;
   rapidjson::SetValueByPointer(dom, "/fixed_speed", 500);
 
   options.set_costing_type(Costing::auto_);
   auto& co = (*options.mutable_costings())[Costing::auto_];
-  sif::ParseBaseCostOptions(*rapidjson::GetValueByPointer(dom, ""), &co, {});
+  sif::ParseBaseCostOptions(*rapidjson::GetValueByPointer(dom, ""), &co, {},
+                            *api.mutable_info()->mutable_warnings());
 
-  ASSERT_EQ(co.options().fixed_speed(), baldr::kDisableFixedSpeed);
+  EXPECT_EQ(co.options().fixed_speed(), baldr::kDisableFixedSpeed);
+  EXPECT_EQ(api.info().warnings().size(), 1);
+  EXPECT_EQ(api.info().warnings(0).code(), 500);
+  EXPECT_THAT(api.info().warnings(0).description(), testing::HasSubstr("fixed_speed"));
 }
 
 TEST_F(FixedSpeedTest, TopAndFixedSpeed) {