+++ valhalla/sif/pedestriancost.h
@@ -19,7 +19,8 @@ namespace sif {
  */
 void ParsePedestrianCostOptions(const rapidjson::Document& doc,
                                 const std::string& costing_options_key,
-                                Costing* pbf_costing);
+                                Costing* pbf_costing,
+                                google::protobuf::RepeatedPtrField<CodedDescription>& warnings);
 
 /**
  * Create a pedestriancost